# backend/openapi-v1.yaml
openapi: 3.0.3
info:
  title: PAMS API
  version: 1.0.0
  description: |
    Permit Assessment & Management System API.
    Phase 1–6: RBAC, pagination, jobs, portal OTP/payments,
    org units, attachments, waterworks module layering, portal confirm dual-write.
servers:
  - url: /api
    description: Current API (treated as v1)
  - url: /api/v1
    description: Explicit v1 alias (same handlers)
...
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: List applications (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: permit_type
          schema: { type: string }
      responses:
        '200':
          description: Paginated application list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApplications'
  /tasks:
    get:
      tags: [Tasks]
      summary: My Work inbox
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Aggregated pending tasks
  /jobs:
    post:
      tags: [Jobs]
      summary: Enqueue PDF generation job
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, applicationId]
              properties:
                type:
                  type: string
                  enum: [assessment_pdf, permit_pdf]
                applicationId:
                  type: string
      responses:
        '202':
          description: Job accepted
...
  /portal-payments:
    get:
      tags: [PortalPayments]
      summary: Staff list of portal payment intents
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: Paginated intents
...
  /org-units:
    get:
      tags: [OrgUnits]
      summary: List organizational units
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Org unit list
  /attachments:
    get:
      tags: [Attachments]
      summary: List attachments for a reference
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: module
          required: true
          schema: { type: string }
        - in: query
          name: reference_type
          required: true
          schema: { type: string }
        - in: query
          name: reference_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Attachment metadata list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    PaginatedApplications:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        pagination:
          $ref: '#/components/schemas/Pagination'
...